Bugfix/elastic updating (#52)
* Added handling errors in try catch for delete method in controller;

* Added refresh true for remove item

+++ src/grainger-items/grainger-items.controller.ts
@@ -181,18 +181,22 @@ export class GraingerItemsController {
     @Body() { itemIds }: { itemIds: string[] },
     @Req() { user }: Request,
   ): Promise<any> {
-    const { result } = await this.itemsService.getAll({
-      user: {
-        name: user.name,
-      },
-      id: In(itemIds),
-    });
-
-    await this.itemsService.delete({
-      id: In(itemIds),
-    });
-
     try {
+      const { result } = await this.itemsService.getAll({
+        user: {
+          name: user.name,
+        },
+        id: In(itemIds),
+      });
+
+      if (!result.length) {
+        throw new Error(`Items doesn't exist`);
+      }
+
+      await this.itemsService.delete({
+        id: In(itemIds),
+      });
+
       await this.itemsSearchService.delete(itemIds);
 
       this.logger.debug(
